feat(PlayerDamage.cs)
데미지 계산 함수 구현

+++ Assets/Player/PlayerOpiton/PlayerStat.cs
@@ -6,7 +6,7 @@ public class Stat : MonoBehaviour
 {
     [SerializeField] public float playerATK; // 플레이어 공격력
     [SerializeField] public float playerSpeed; // 이동 속도
-    [SerializeField] public float playerHealth; // 플레이어 체력
+    [SerializeField] public int playerHealth; // 플레이어 체력 (절반 하트 당 1로 계산)
     [SerializeField] public float playerDefence; // 플레이어 방어력
 }
